expose ability to create topics (#67)
